HP patch for initializing select plugin before setting bitmaps in slurmctld and changing state to cancelled for jobs ended by ctrl-c

+++ src/slurmctld/read_config.c
@@ -702,7 +702,17 @@ int read_slurm_conf(int recover)
 		}
 		reset_first_job_id();
 	}
-	reset_job_bitmaps();
+
+	if ((select_g_node_init(node_record_table_ptr, node_record_count)
+			!= SLURM_SUCCESS) 
+	    || (select_g_block_init(part_list) != SLURM_SUCCESS) 
+	    || (select_g_job_init(job_list) != SLURM_SUCCESS)) {
+		error("failed to initialize node selection plugin state");
+		abort();
+	}
+
+	reset_job_bitmaps();		/* must follow select_g_job_init() */
+
 	(void) _sync_nodes_to_jobs();
 	(void) sync_job_files();
 	_purge_old_node_state(old_node_table_ptr, old_node_record_count);
@@ -712,14 +722,7 @@ int read_slurm_conf(int recover)
 #ifdef 	HAVE_ELAN
 	_validate_node_proc_count();
 #endif
-	if ((select_g_node_init(node_record_table_ptr, node_record_count)
-			!= SLURM_SUCCESS) 
-	|| (select_g_block_init(part_list) != SLURM_SUCCESS) 
-	|| (select_g_job_init(job_list) != SLURM_SUCCESS)) {
-		error("failed to initialize node selection plugin state");
-		abort();
-	}
-	(void) _sync_nodes_to_comp_job(); /* must follow select_g_node_init() */
+	(void) _sync_nodes_to_comp_job();/* must follow select_g_node_init() */
 	load_part_uid_allow_list(1);
 
 	/* sort config_list by weight for scheduling */
